Add Hash Resolution
This is part of a big commit set that adds resolution of FNV1a hashes to
the dispatch table API.

Basically, this allows a client library to expose its own function
resolution interface without creating a compile-time dependency on this
library. Instead of having to maintain a parallel list of command
indices, a client library can simply compute the FNV1a hash of the
Vulkan command name. Since the complete list of Vulkan commands in this
library might differ from what a client expects, the hash API can return
nullptr.

Direct clients of this library should still prefer the index based api
which is simpler and doesn't permit a null return value (even if the
pointer-to value can be null).

In addition to the hash stuff, this commit reorganizes a lot of the
template structure. The Vulkan command list is now generated in
defs.inl.in as an X-macro. This makes it simpler to reuse the list for
both command enumerations and hash resolution methods. dispatch::error
is also in defs.inl.in now.

Finally, and as usual, this commit has some clean up of the templates
and documentation.