Reduce Generated Includes
This reduces the overall number of generated header files from 3 to 1.
Only defs.inl.in remains. The other header files are now static and
provided as normal.

To achieve this, a few macros are introduced (for array sizes) and the
specification_version() method is broken out to a free function in
defs.inl. specification_version() is still a constant expression in any
case.