Add to_command Functions
This commit adds functions for mapping hash values to command indices
without retrieving any functions from a table. Potentially, this
provides a large speed-up for clients that need to compute hashes at
run-time. As long as they compute the hash and resolve the command index
ahead of resolving the function pointer.