Update Build Scripts
This commit does two things. First, it changes all the test names
because I am a perfectionist. As a result, I feel the need to
obsessively normalize these names after the fact. Second, it changes the
test build script so that the additional test dependencies are only
resolved once. Again, this is because I am too neurotic to allow it to
check twice for libvulkan.