Add Instance-Level Device Table to Device-Level Device Table Extension
This adds a constructor that will extend an instance-level device
dispatch table into a device-level device dispatch table. Essentially,
this involves reloading all of the function pointers (costly) from the
new device handle.

This is basically equivalent to calling:

table{ gdt, instance::table{ gdt, instance }, device }

but with the side effect of requiring a bunch of extra function
resolutions.

I wouldn't recommend actually doing this unless you need to, but I think
the functionality should be there if instance-level device tables are
supported.