Add Preconditions and Postconditions
This changes a few things. First, it adds Megatech Assertions as a
submodule. This is to bring in the MEGATECH_PRECONDITION and
MEGATECH_POSTCONDITION macros. Second, it bumps the version to 1.0.1.
Third, it does a little formatting clean up. Finally, it updates the
build script to include the new dependency.

The Megatech Assertions dependency is only meaningful in debug builds.
In release builds, only a partial dependency on the header file is used.
The linker should ellide the library anyway in release configurations,
but this ensures that.